@@ -4,7 +4,7 @@ description: Demonstrates how to use the flutter_pdfview plugin.
 publish_to: "none"
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -12,8 +12,8 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  path_provider: ^1.2.0
+  cupertino_icons: ^1.0.2
+  path_provider: ^2.0.1
 
 dev_dependencies:
   flutter_test:
