@@ -4,7 +4,7 @@ version: 1.0.4
 homepage: https://github.com/endigo/flutter_pdfview
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.13+hotfix.5"
 
 dependencies:
