@@ -7,17 +7,17 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
 typedef PDFViewCreatedCallback = void Function(PDFViewController controller);
-typedef RenderCallback = void Function(int pages);
-typedef PageChangedCallback = void Function(int page, int total);
+typedef RenderCallback = void Function(int? pages);
+typedef PageChangedCallback = void Function(int? page, int? total);
 typedef ErrorCallback = void Function(dynamic error);
-typedef PageErrorCallback = void Function(int page, dynamic error);
-typedef LinkHandlerCallback = void Function(String uri);
+typedef PageErrorCallback = void Function(int? page, dynamic error);
+typedef LinkHandlerCallback = void Function(String? uri);
 
 enum FitPolicy { WIDTH, HEIGHT, BOTH }
 
 class PDFView extends StatefulWidget {
   const PDFView({
-    Key key,
+    Key? key,
     this.filePath,
     this.pdfData,
     this.onViewCreated,
@@ -45,12 +45,12 @@ class PDFView extends StatefulWidget {
   _PDFViewState createState() => _PDFViewState();
 
   /// If not null invoked once the web view is created.
-  final PDFViewCreatedCallback onViewCreated;
-  final RenderCallback onRender;
-  final PageChangedCallback onPageChanged;
-  final ErrorCallback onError;
-  final PageErrorCallback onPageError;
-  final LinkHandlerCallback onLinkHandler;
+  final PDFViewCreatedCallback? onViewCreated;
+  final RenderCallback? onRender;
+  final PageChangedCallback? onPageChanged;
+  final ErrorCallback? onError;
+  final PageErrorCallback? onPageError;
+  final LinkHandlerCallback? onLinkHandler;
 
   /// Which gestures should be consumed by the pdf view.
   ///
@@ -61,15 +61,15 @@ class PDFView extends StatefulWidget {
   ///
   /// When this set is empty or null, the pdf view will only handle pointer events for gestures that
   /// were not claimed by any other gesture recognizer.
-  final Set<Factory<OneSequenceGestureRecognizer>> gestureRecognizers;
+  final Set<Factory<OneSequenceGestureRecognizer>>? gestureRecognizers;
 
   /// The initial URL to load.
-  final String filePath;
-  final Uint8List pdfData;
+  final String? filePath;
+  final Uint8List? pdfData;
 
   final bool enableSwipe;
   final bool swipeHorizontal;
-  final String password;
+  final String? password;
   final bool nightMode;
   final bool autoSpacing;
   final bool pageFling;
@@ -110,7 +110,7 @@ class _PDFViewState extends State<PDFView> {
     final PDFViewController controller = PDFViewController._(id, widget);
     _controller.complete(controller);
     if (widget.onViewCreated != null) {
-      widget.onViewCreated(controller);
+      widget.onViewCreated!(controller);
     }
   }
 
@@ -137,10 +137,10 @@ class _CreationParams {
     );
   }
 
-  final String filePath;
-  final Uint8List pdfData;
+  final String? filePath;
+  final Uint8List? pdfData;
 
-  final _PDFViewSettings settings;
+  final _PDFViewSettings? settings;
 
   Map<String, dynamic> toMap() {
     Map<String, dynamic> params = {
@@ -148,7 +148,7 @@ class _CreationParams {
       'pdfData': pdfData,
     };
 
-    params.addAll(settings.toMap());
+    params.addAll(settings!.toMap());
 
     return params;
   }
@@ -182,17 +182,17 @@ class _PDFViewSettings {
         preventLinkNavigation: widget.preventLinkNavigation);
   }
 
-  final bool enableSwipe;
-  final bool swipeHorizontal;
-  final String password;
-  final bool nightMode;
-  final bool autoSpacing;
-  final bool pageFling;
-  final bool pageSnap;
-  final int defaultPage;
-  final FitPolicy fitPolicy;
-  final bool fitEachPage;
-  final bool preventLinkNavigation;
+  final bool? enableSwipe;
+  final bool? swipeHorizontal;
+  final String? password;
+  final bool? nightMode;
+  final bool? autoSpacing;
+  final bool? pageFling;
+  final bool? pageSnap;
+  final int? defaultPage;
+  final FitPolicy? fitPolicy;
+  final bool? fitEachPage;
+  final bool? preventLinkNavigation;
 
   Map<String, dynamic> toMap() {
     return <String, dynamic>{
@@ -239,40 +239,40 @@ class PDFViewController {
 
   final MethodChannel _channel;
 
-  _PDFViewSettings _settings;
+  late _PDFViewSettings _settings;
 
   PDFView _widget;
 
-  Future<bool> _onMethodCall(MethodCall call) async {
+  Future<bool?> _onMethodCall(MethodCall call) async {
     switch (call.method) {
       case 'onRender':
         if (_widget.onRender != null) {
-          _widget.onRender(call.arguments['pages']);
+          _widget.onRender!(call.arguments['pages']);
         }
 
         return null;
       case 'onPageChanged':
         if (_widget.onPageChanged != null) {
-          _widget.onPageChanged(
+          _widget.onPageChanged!(
               call.arguments['page'], call.arguments['total']);
         }
 
         return null;
       case 'onError':
         if (_widget.onError != null) {
-          _widget.onError(call.arguments['error']);
+          _widget.onError!(call.arguments['error']);
         }
 
         return null;
       case 'onPageError':
         if (_widget.onPageError != null) {
-          _widget.onPageError(call.arguments['page'], call.arguments['error']);
+          _widget.onPageError!(call.arguments['page'], call.arguments['error']);
         }
 
         return null;
       case 'onLinkHandler':
         if (_widget.onLinkHandler != null) {
-          _widget.onLinkHandler(call.arguments);
+          _widget.onLinkHandler!(call.arguments);
         }
 
         return null;
@@ -281,18 +281,18 @@ class PDFViewController {
         '${call.method} was invoked but has no handler');
   }
 
-  Future<int> getPageCount() async {
-    final int pageCount = await _channel.invokeMethod('pageCount');
+  Future<int?> getPageCount() async {
+    final int? pageCount = await _channel.invokeMethod('pageCount');
     return pageCount;
   }
 
-  Future<int> getCurrentPage() async {
-    final int currentPage = await _channel.invokeMethod('currentPage');
+  Future<int?> getCurrentPage() async {
+    final int? currentPage = await _channel.invokeMethod('currentPage');
     return currentPage;
   }
 
-  Future<bool> setPage(int page) async {
-    final bool isSet = await _channel.invokeMethod('setPage', <String, dynamic>{
+  Future<bool?> setPage(int page) async {
+    final bool? isSet = await _channel.invokeMethod('setPage', <String, dynamic>{
       'page': page,
     });
     return isSet;
